✨ feat(version): add API version
