+++ .goreleaser.yaml
@@ -23,7 +23,7 @@ builds:
       - windows
       - darwin
     ldflags:
-        - "-s -w -X {{ .ModulePath }}/cmd.Version={{.Version}}"
+        - "-s -w -X {{ .ModulePath }}/pkg/version.Version=v{{.Version}}"
 
 archives:
   - formats: [binary]
