+++ cmd/root.go
@@ -6,14 +6,10 @@ SPDX-License-Identifier: BSD-3-Clause
 package cmd
 
 import (
-	"context"
 	"os"
-	"time"
 
 	"github.com/fatih/color"
-	"github.com/mattn/go-isatty"
-	"github.com/outscale/gli/pkg/errors"
-	"github.com/outscale/gli/pkg/update"
+	"github.com/outscale/gli/cmd/prerun"
 	"github.com/spf13/cobra"
 )
 
@@ -41,21 +37,7 @@ var rootCmd = &cobra.Command{
 `) + d(`one CLI to bring them all
 `) + e(`and in the terminal bind them.`),
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-		if !isatty.IsTerminal(os.Stdout.Fd()) {
-			return
-		}
-		ctx := context.Background()
-		ghCtx, cancel := context.WithTimeout(ctx, time.Second)
-		defer cancel()
-		latest, err := update.LatestRelease(ghCtx)
-		if err != nil {
-			errors.Warn("❌ Unable to fetch latest release: %v\n", err)
-			return
-		}
-		if latest == "" || latest == Version {
-			return
-		}
-		_, _ = color.New(color.FgGreen).Add(color.Bold).Printf("⬆️ New version %s detected - call gli update to update\n", latest)
+		prerun.CheckUpdate(cmd, args)
 	},
 }
