+++ pkg/version/version.go
@@ -1,0 +1,3 @@
+package version
+
+var Version = "v0.0.0-dev"
